@@ -23,14 +23,14 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "13.3.11",
-    "@angular/cdk": "13.3.9",
-    "@angular/common": "13.3.11",
-    "@angular/core": "13.3.11",
-    "@angular/forms": "13.3.11",
-    "@angular/platform-browser": "13.3.11",
-    "@angular/platform-browser-dynamic": "13.3.11",
-    "@angular/router": "13.3.11",
+    "@angular/animations": "14.0.0",
+    "@angular/cdk": "14.0.0",
+    "@angular/common": "14.0.0",
+    "@angular/core": "14.0.0",
+    "@angular/forms": "14.0.0",
+    "@angular/platform-browser": "14.0.0",
+    "@angular/platform-browser-dynamic": "14.0.0",
+    "@angular/router": "14.0.0",
     "@asymmetrik/ngx-leaflet": "13.0.2",
     "@capacitor-community/fcm": "2.0.2",
     "@capacitor-community/firebase-analytics": "1.0.1",
@@ -48,6 +48,7 @@
     "@capacitor/splash-screen": "1.2.2",
     "@capacitor/status-bar": "1.0.8",
     "@capacitor/storage": "1.2.5",
+    "@ionic-native/core": "^5.36.0",
     "@ionic-native/in-app-purchase-2": "5.36.0",
     "@ionic-super-tabs/angular": "7.0.8",
     "@ionic/angular": "6.1.9",
@@ -88,9 +89,9 @@
     "@angular-eslint/schematics": "13.2.1",
     "@angular-eslint/template-parser": "13.2.1",
     "@angular/cli": "14.0.0",
-    "@angular/compiler": "13.3.11",
-    "@angular/compiler-cli": "13.3.11",
-    "@angular/language-service": "13.3.11",
+    "@angular/compiler": "14.0.0",
+    "@angular/compiler-cli": "14.0.0",
+    "@angular/language-service": "14.0.0",
     "@capacitor/cli": "3.5.1",
     "@cypress/schematic": "2.0.0",
     "@ionic/angular-toolkit": "6.1.0",
