@@ -1,5 +1,5 @@
 import {AfterViewInit, Component, Input, OnInit, ViewChild} from '@angular/core';
-import {FormControl} from '@angular/forms';
+import {FormControl, UntypedFormControl} from '@angular/forms';
 import {Observable, of} from 'rxjs';
 import {MemberEntry} from '../../../core/api/models/member-entry';
 import {MembersService} from '../../../core/api/services/members.service';
@@ -23,7 +23,7 @@ export class SearchMemberComponent implements OnInit, AfterViewInit {
 
     @ViewChild('searchbar') searchBar: any;
 
-    searchBox: FormControl;
+    searchBox: FormControl<string>;
     members$: Observable<MemberEntry[]>;
     loading = false;
     skeletonRows: number[];
@@ -40,7 +40,7 @@ export class SearchMemberComponent implements OnInit, AfterViewInit {
 
     ngOnInit(): void {
         this.skeletonRows = [...Array(16).keys()].map(() => Math.floor(Math.random() * 40) + 30);
-        this.searchBox = new FormControl('');
+        this.searchBox = new FormControl<string>('');
         this.members$ = this.searchBox.valueChanges.pipe(
             filter((text) => text?.length >= 3 || text.length === 0),
             debounceTime(500),
